Revert "Improve the WordPress developer console with a few new features."
This reverts commit 4655ea2dcd9f7bebcbbe37c2e225d3f0d2410774.
