+++ src/components/results/index.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import classnames from 'classnames';
-import { getResults } from '../../state/results/selectors';
-import Result from './response';
+import JSONTree from 'react-json-tree';
 
 import './style.css';
+
+import RequestHeader from './header';
+import { getResults } from '../../state/results/selectors';
+import { escapeLikeJSON, stringify } from './utils';
 
 class Results extends React.Component {
 	componentWillMount() {
@@ -24,15 +27,65 @@ class Results extends React.Component {
 	render() {
 		const { results } = this.props;
 
+		const jsonTheme = {
+			extend: 'default',
+			tree: {
+				MozUserSelect: 'text',
+				WebkitUserSelect: 'text',
+			},
+			nestedNodeItemString: ( node, expandedNodes, type, expanded ) => {
+				return {
+					className: expanded ? 'expanded' : 'collapsed',
+				};
+			},
+		};
+
+		const expandFalse = () => false;
+		const getItemString = ( type, data, itemType, itemString ) => (
+			/* eslint-disable react/no-danger */
+			<span
+				className="collapsed-content"
+				dangerouslySetInnerHTML={ { __html: `${ itemString } <span class="content">${ stringify( data ) }</span>` } }
+			/>
+			/* eslint-enable react/no-danger */
+		);
+
+		const valueRenderer = value => {
+			if ( typeof value === 'string' ) {
+				const valueWithoutQuotes = value.replace( /^"|"$/g, '' );
+				let valueEscaped;
+				if ( valueWithoutQuotes === value ) {
+					// Probably a boolean or number
+					valueEscaped = escapeLikeJSON( value );
+				} else {
+					// String
+					valueEscaped = '"' + escapeLikeJSON( valueWithoutQuotes ) + '"';
+				}
+				return (
+					<span className="expanded-value">{ valueEscaped }</span>
+				);
+			}
+
+			return value;
+		};
+
 		return (
 			<div className="results">
-				{ results.map( result =>
-					<Result
-						key={ result.id }
-						result={ result }
-						className={ classnames( 'request', { error: result.response && !! result.response.error } ) }
-
-					/>
+				{ results.map( ( result, index ) =>
+					<div key={ result.id } className={ classnames( 'request', { error: result.response && !! result.response.error } ) }>
+						<RequestHeader result={ result } />
+						{ result.response && result.response.body &&
+							<div className="response">
+								<JSONTree
+									theme={ jsonTheme }
+									data={ result.response.body }
+									shouldExpandNode={ expandFalse }
+									getItemString={ getItemString }
+									valueRenderer={ valueRenderer }
+								/>
+							</div>
+						}
+					</div>
 				)}
 			</div>
 		);
